+++ adapters/VerizonMedia/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Verizon Media iOS Mediation Adapter Changelog
 
+#### Version 1.2.1.0
+- Added support for Verizon Ads Standard Edition 1.2.1
+- Added support for native ads
+- Added support for rewarded video ads
+
 #### Version 1.1.2.0
 - Initial release.
 - Verified compatibility with Verizon Ads Standard Edition version 1.1.2.